fix(input): бесфокусный ввод MoTeC-style и атомарный Bulk-пуш в UART
- Реализован InputController для изоляции клавиатурного ввода от файлов окон (WPF DataContext).
- Ослеплены шаблоны TextBox (OneWay, Focusable=False) для предотвращения прыжков каретки и ложного приписывания нулей.
- Внедрен временный InputBuffer в VariableViewModel.Core для стерильного накопления строк.
- Устранена гонка потоков и пулемет левых Bulk-пакетов при вводе цифр за счет полной очистки сеттеров ячеек MatrixCellViewModel.
- Реализован атомарный выстрел NetworkCommand по нажатию Enter строго из верхнеуровневого драйвера ввода.
- Восстановлена циклическая навигация по Tab строго по калибровочным параметрам (IsParam).
- Реализована функция "Jump to Working Point" по нажатию Пробела для мгновенного магнитного прыжка на зеленую режимную ячейку.
- Оживлена кнопка "СОХРАНИТЬ ВО FLASH", отправляющая системный триггер-кадр AA-00-03-00-00 в плату BlackPill.

+++ WpfCalibratorV3.0/ViewModels/MainViewModel/MainViewModel.Core.cs
@@ -179,12 +179,12 @@ public partial class MainViewModel : INotifyPropertyChanged
         {
             if (param.IsLutLinked)
             {
-                var axisXValues = param.BoundAxisX!.MatrixData.Cast<float>().ToArray();
-                var axisYValues = param.BoundAxisY!.MatrixData.Cast<float>().ToArray();
+                var axisXValues = param.BoundAxisX!.MatrixData.Cast<double>().ToArray();
+                var axisYValues = param.BoundAxisY!.MatrixData.Cast<double>().ToArray();
 
                 param.CalculateWorkingPoint(
-                    (float)param.BoundInputX!.CurrentValue,
-                    (float)param.BoundInputY!.CurrentValue,
+                    param.BoundInputX!.CurrentValue,
+                    param.BoundInputY!.CurrentValue,
                     axisXValues,
                     axisYValues
                 );
